# backend/openapi/api.yaml
openapi: 3.1.0
info:
    description: Executable HTTP contract for the supported Synaudio V1 surface.
    title: Synaudio V1 API
    version: 1.0.0
servers:
    - url: http://localhost:8080
paths:
...
    /api/v1/admin/creative-decisions/{decisionID}/postpone:
        parameters:
            - in: path
              name: decisionID
              required: true
              schema:
                type: string
        post:
            operationId: postApiV1AdminCreativeDecisionsDecisionidPostpone
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PostponeCreativeDecisionRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreativeDecisionPascal'
                    description: Creative decision postponed
                "400":
                    $ref: '#/components/responses/BadRequest'
                "401":
                    $ref: '#/components/responses/Unauthenticated'
                "403":
                    $ref: '#/components/responses/Forbidden'
            security:
                - bearerAuth: []
...
    /api/v1/admin/stories/{storyID}/arcs:
        get:
            operationId: getApiV1AdminStoriesStoryidArcs
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StoryArcListResponse'
                    description: Story arcs
                "400":
                    $ref: '#/components/responses/BadRequest'
                "401":
                    $ref: '#/components/responses/Unauthenticated'
                "403":
                    $ref: '#/components/responses/Forbidden'
            security:
                - bearerAuth: []
        parameters:
            - in: path
              name: storyID
              required: true
              schema:
                type: string
        post:
            operationId: postApiV1AdminStoriesStoryidArcs
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateArcRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StoryArcPascal'
                    description: Story arc created
                "400":
                    $ref: '#/components/responses/BadRequest'
                "401":
                    $ref: '#/components/responses/Unauthenticated'
                "403":
                    $ref: '#/components/responses/Forbidden'
            security:
                - bearerAuth: []
...
    /api/v1/admin/stories/{storyID}/bible/versions:
        parameters:
            - in: path
              name: storyID
              required: true
              schema:
                type: string
        post:
            operationId: postApiV1AdminStoriesStoryidBibleVersions
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/VersionContentRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StoryBibleVersionPascal'
                    description: Story bible version created
                "400":
                    $ref: '#/components/responses/BadRequest'
                "401":
                    $ref: '#/components/responses/Unauthenticated'
                "403":
                    $ref: '#/components/responses/Forbidden'
            security:
                - bearerAuth: []
...
    /api/v1/admin/stories/{storyID}/canon-branches/active-official:
        get:
            operationId: getApiV1AdminStoriesStoryidCanonBranchesActiveOfficial
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CanonBranchPascal'
                    description: Active OFFICIAL canon branch
            security:
                - bearerAuth: []
        parameters:
            - in: path
              name: storyID
              required: true
              schema:
                type: string
...
    /api/v1/admin/stories/{storyID}/characters:
        get:
            operationId: getApiV1AdminStoriesStoryidCharacters
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CharacterListResponse'
                    description: Story characters
                "400":
                    $ref: '#/components/responses/BadRequest'
                "401":
                    $ref: '#/components/responses/Unauthenticated'
                "403":
                    $ref: '#/components/responses/Forbidden'
            security:
                - bearerAuth: []
        parameters:
            - in: path
              name: storyID
              required: true
              schema:
                type: string
        post:
            operationId: postApiV1AdminStoriesStoryidCharacters
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateCharacterRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CharacterPascal'
                    description: Character created
                "400":
                    $ref: '#/components/responses/BadRequest'
                "401":
                    $ref: '#/components/responses/Unauthenticated'
                "403":
                    $ref: '#/components/responses/Forbidden'
            security:
                - bearerAuth: []
...
    /api/v1/admin/stories/{storyID}/ending/versions:
        parameters:
            - in: path
              name: storyID
              required: true
              schema:
                type: string
        post:
            operationId: postApiV1AdminStoriesStoryidEndingVersions
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/VersionContentRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EndingPlanVersionPascal'
                    description: Ending plan version created
                "400":
                    $ref: '#/components/responses/BadRequest'
                "401":
                    $ref: '#/components/responses/Unauthenticated'
                "403":
                    $ref: '#/components/responses/Forbidden'
            security:
                - bearerAuth: []
...
components:
    responses:
        BadRequest:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
            description: Request validation failed
        Forbidden:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
            description: Authenticated principal is not allowed to perform the operation
        Unauthenticated:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
            description: Authentication required or credential invalid
    schemas:
...
        CanonBranchPascal:
            properties:
                ID:
                    type: string
                Status:
                    type: string
                StoryID:
                    type: string
                Type:
                    type: string
            type: object
...
        CharacterListResponse:
            properties:
                characters:
                    items:
                        $ref: '#/components/schemas/CharacterPascal'
                    type: array
            required:
                - characters
            type: object
        CharacterPascal:
            properties:
                CanonicalName:
                    type: string
                CurrentProfileVersionID:
                    type: string
                ID:
                    type: string
                Importance:
                    type: string
                StoryID:
                    type: string
            type: object
...
        CreateArcRequest:
            properties:
                content:
                    additionalProperties: true
                    type: object
            required:
                - content
            type: object
...
        CreateCharacterRequest:
            properties:
                importance:
                    type: string
                name:
                    type: string
                profile:
                    additionalProperties: true
                    type: object
            required:
                - name
            type: object
...
        CreativeDecisionPascal:
            properties:
                ArcID:
                    type: string
                BlockingLevel:
                    type: string
                ChapterID:
                    type: string
                ContextSummary:
                    type: string
                DecisionType:
                    type: string
                ID:
                    type: string
                Origin:
                    type: string
                Question:
                    type: string
                Severity:
                    type: string
                Status:
                    type: string
                StoryID:
                    type: string
            type: object
...
        EndingPlanVersionPascal:
            properties:
                BasedOnVersionID:
                    type: string
                Content:
                    additionalProperties: true
                    type: object
                CreatedBy:
                    type: string
                ID:
                    type: string
                StoryID:
                    type: string
                VersionNo:
                    type: integer
            type: object
        ErrorBody:
            properties:
                code:
                    type: string
                message:
                    type: string
            required:
                - code
                - message
            type: object
        ErrorResponse:
            properties:
                error:
                    $ref: '#/components/schemas/ErrorBody'
            required:
                - error
            type: object
...
        PostponeCreativeDecisionRequest:
            properties:
                reason:
                    maxLength: 500
                    minLength: 1
                    type: string
            required:
                - reason
            type: object
...
        StoryArcListResponse:
            properties:
                arcs:
                    items:
                        $ref: '#/components/schemas/StoryArcPascal'
                    type: array
            required:
                - arcs
            type: object
        StoryArcPascal:
            properties:
                CurrentVersionID:
                    type: string
                ID:
                    type: string
                Ordinal:
                    type: integer
                Status:
                    type: string
                StoryID:
                    type: string
            type: object
...
        StoryBibleVersionPascal:
            properties:
                BasedOnVersionID:
                    type: string
                Content:
                    additionalProperties: true
                    type: object
                CreatedBy:
                    type: string
                ID:
                    type: string
                StoryID:
                    type: string
                VersionNo:
                    type: integer
            type: object
...
        VersionContentRequest:
            properties:
                content:
                    additionalProperties: true
                    type: object
            required:
                - content
            type: object
...
    securitySchemes:
        bearerAuth:
            bearerFormat: JWT
            scheme: bearer
            type: http